+++ app/src/main/kotlin/ac/mdiq/podcini/ui/dialog/CustomFeedNameDialog.kt
@@ -30,9 +30,10 @@ class CustomFeedNameDialog(activity: Activity, private var feed: Feed) {
             .setTitle(R.string.rename_feed_label)
             .setPositiveButton(android.R.string.ok) { _: DialogInterface?, _: Int ->
                 val newTitle = binding.editText.text.toString()
-                feed = unmanaged(feed)
-                feed.setCustomTitle1(newTitle)
-                feed = upsertBlk(feed) {}
+//                feed = unmanaged(feed)
+                feed = upsertBlk(feed) {
+                    it.setCustomTitle1(newTitle)
+                }
             }
             .setNeutralButton(R.string.reset, null)
             .setNegativeButton(R.string.cancel_label, null)